credit-fraud-detector/requirements.txt
aiobotocore==2.4.2
aiofiles==24.1.0
aiohttp==3.9.5
aioitertools==0.11.0
aiomysql==0.2.0
aiosignal==1.3.1
alembic==1.13.1
altair==4.2.2
aniso8601==9.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
avro==1.11.3
Babel==2.15.0
backoff==2.2.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
bleach==6.1.0
blinker==1.8.2
boto3==1.34.106
botocore==1.34.113
build==1.2.1
cachetools==5.3.3
category-encoders==2.6.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
commonmark==0.9.1
confluent-kafka==2.3.0
contourpy==1.2.1
cookiecutter==2.6.0
cryptography==42.0.8
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dnspython==2.6.1
docker==7.1.0
dynaconf==3.2.5
email_validator==2.2.0
entrypoints==0.4
evidently==0.4.25
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastavro==1.8.4
fastjsonschema==2.20.0
FLAML==1.2.4
Flask==3.0.3
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2022.11.0
furl==2.1.3
future==1.0.0
gcsfs==2022.11.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
great-expectations==0.18.12
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
hopsworks==3.7.0
hsfs==3.7.6
hsml==3.7.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
javaobj-py3==0.4.4
jedi==0.19.1
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kaleido==0.2.1
kedro==0.19.6
kedro-datasets==3.0.1
kedro-mlflow==0.12.2
kedro-viz==9.1.0
kiwisolver==1.4.5
lazy_loader==0.4
lightgbm==3.3.5
llvmlite==0.43.0
makefun==1.15.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mlflow==2.14.1
mock==5.1.0
more-itertools==10.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nannyml==0.10.6
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
numba==0.60.0
numpy==1.24.4
nvidia-nccl-cu12==2.22.3
oauthlib==3.2.2
omegaconf==2.3.0
opensearch-py==2.4.2
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
orderedmultidict==1.0.1
orjson==3.10.5
overrides==7.7.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.5.0
pre-commit-hooks==4.6.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pyfiglet==0.8.post1
Pygments==2.18.0
PyHopsHive==0.6.4.1.dev0
pyhumps==1.6.1
pyjks==20.0.0
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.1.2
pyproject_hooks==1.1.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
querystring-parser==1.2.4
referencing==0.35.1
regex==2024.5.10
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rope==1.13.0
rpds-py==0.18.1
rsa==4.9
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.8
s3fs==2022.11.0
s3transfer==0.10.2
scikit-learn==1.3.2
scipy==1.14.0
seaborn==0.13.2
secure==0.3.0
segment-analytics-python==2.3.2
Send2Trash==1.8.3
shap==0.46.0
shellingham==1.5.4
six==1.16.0
slicer==0.0.8
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.15
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.2
strawberry-graphql==0.235.0
streamlit==1.36.0
tenacity==8.4.2
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
thrift==0.20.0
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
toposort==1.10
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
twofish==0.3.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
types-PyYAML==6.0.12.20240311
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
uri-template==1.3.0
urllib3==1.26.19
uvicorn==0.29.0
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
watchgod==0.8.2
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
xgboost==2.1.0
yarl==1.9.4
zipp==3.19.2
